refactor: consolidate parser shared imports
