+++ src/parser/src/gml-ast-builder.js
@@ -1,6 +1,6 @@
 import GameMakerLanguageParserVisitor from "./generated/GameMakerLanguageParserVisitor.js";
-import { getLineBreakCount } from "./shared/utils/line-breaks.js";
-import { getNonEmptyTrimmedString } from "./shared/utils.js";
+import { getLineBreakCount } from "../../shared/utils/line-breaks.js";
+import { getNonEmptyTrimmedString } from "../../shared/utils.js";
 import ScopeTracker from "./scope-tracker.js";
 import { ScopeOverrideKeyword } from "./scope-override-keywords.js";
 import BinaryExpressionDelegate from "./binary-expression-delegate.js";
