+++ src/parser/src/gml-parser.js
@@ -6,10 +6,10 @@ import GameMakerASTBuilder from "./gml-ast-builder.js";
 import GameMakerParseErrorListener, {
     GameMakerLexerErrorListener
 } from "./gml-syntax-error.js";
-import { isObjectLike } from "./shared/object-utils.js";
-import { isErrorLike } from "./shared/utils/capability-probes.js";
-import { getLineBreakCount } from "./shared/utils/line-breaks.js";
-import { enqueueObjectChildValues } from "./shared/ast.js";
+import { enqueueObjectChildValues } from "../../shared/ast.js";
+import { isObjectLike } from "../../shared/object-utils.js";
+import { isErrorLike } from "../../shared/utils/capability-probes.js";
+import { getLineBreakCount } from "../../shared/utils/line-breaks.js";
 
 function normalizeSimpleEscapeCase(text) {
     if (typeof text !== "string" || text.length === 0) {
@@ -422,4 +422,4 @@ export default class GMLParser {
     }
 }
 
-export { getLineBreakCount } from "./shared/utils/line-breaks.js";
+export { getLineBreakCount } from "../../shared/utils/line-breaks.js";
