+++ src/parser/src/scope-tracker.js
@@ -1,5 +1,5 @@
-import { assignClonedLocation } from "./shared/ast.js";
-import { isObjectLike, toArray } from "./shared/utils.js";
+import { assignClonedLocation } from "../../shared/ast.js";
+import { isObjectLike, toArray } from "../../shared/utils.js";
 import {
     ScopeOverrideKeyword,
     formatKnownScopeOverrideKeywords,
@@ -158,7 +158,7 @@ export default class ScopeTracker {
      * grammar branches without bespoke checks.
      *
      * @param {string | null | undefined} name Identifier being declared.
-     * @param {import("./shared/ast.js").GameMakerAstNode | null | undefined} node
+     * @param {import("../../shared/ast.js").GameMakerAstNode | null | undefined} node
      *        AST node representing the declaration site. The node is mutated to
      *        include scope and classification metadata when provided.
      * @param {{ scopeOverride?: unknown, tags?: Iterable<string>, kind?: string }}
