--- src/parser/src/shared/ast.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/ast.js";
