--- src/parser/src/shared/object-utils.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/object-utils.js";
