--- src/parser/src/shared/utils/line-breaks.js
@@ -1,1 +1,0 @@
-export * from "../../../../shared/utils/line-breaks.js";
