+++ src/parser/src/utils/recognition-exception.js
@@ -1,5 +1,8 @@
-import { hasFunction, isErrorLike } from "./capability-probes.js";
-import { isObjectLike } from "../object-utils.js";
+import {
+    hasFunction,
+    isErrorLike
+} from "../../../shared/utils/capability-probes.js";
+import { isObjectLike } from "../../../shared/object-utils.js";
 
 function hasOffendingTokenProbe(value) {
     if (value?.offendingToken !== undefined) {
